+++ packages/variant-quickjs-singlefile-mjs-release-sync/src/index.ts
@@ -8,6 +8,7 @@ import type { QuickJSSyncVariant } from "@jitl/quickjs-ffi-types"
  *
  * | Variable            |    Setting                     |    Description    |
  * | --                  | --                             | --                |
+ * | library             | quickjs             | The original [bellard/quickjs](https://github.com/bellard/quickjs) library. Version [2024-01-13+229b07b9](https://github.com/bellard/quickjs/commit/229b07b9b2c811eaf84db209a1d6f9e2a8a7b0d9) vendored to quickjs-emscripten on 2024-02-11. |
  * | releaseMode         | release         | Optimized for performance; use when building/deploying your application. |
  * | syncMode            | sync            | The default, normal build. Note that both variants support regular async functions. |
  * | emscriptenInclusion | singlefile | The WASM runtime is included directly in the JS file. Use if you run into issues with missing .wasm files when building or deploying your app. |