+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.28.0 (unreleased)
+
+- [#155] Update library versions and add versions to documentation.
+  - quickjs version [2024-01-13+229b07b9](https://github.com/bellard/quickjs/commit/229b07b9b2c811eaf84db209a1d6f9e2a8a7b0d9) vendored to quickjs-emscripten on 2024-02-11.
+    - Evaluating a ES Module with `context.evalCode(...)` now returns a Promise object instead of `undefined`.
+  - quickjs-ng version [git+229b07b9](https://github.com/quickjs-ng/quickjs/commit/229b07b9b2c811eaf84db209a1d6f9e2a8a7b0d9) vendored to quickjs-emscripten on 2024-02-11.
+
 ## v0.27.0
 
 - [#147] Support providing and retrieving WebAssembly.Memory